use bottom navbar in notifications
